requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
numpy==1.26.1
scipy==1.11.3
SQLAlchemy==2.0.23
python-dateutil==2.8.2
alembic==1.12.1
aiofiles==23.2.1 